// src/types/subsquid/test/v62.ts
export interface AssetId32 {
    code: Bytes
}
// ...
export type AccountId32 = Bytes
// ...
export interface SwapBatchInfo {
    outcomeAssetId: AssetId32
    outcomeAssetReuse: bigint
    dexId: number
    receivers: BatchReceiverInfo[]
}

export interface BatchReceiverInfo {
    accountId: AccountId32
    targetAmount: bigint
}

export type FilterMode = FilterMode_AllowSelected | FilterMode_Disabled | FilterMode_ForbidSelected

export interface FilterMode_AllowSelected {
    __kind: 'AllowSelected'
}

export interface FilterMode_Disabled {
    __kind: 'Disabled'
}

export interface FilterMode_ForbidSelected {
    __kind: 'ForbidSelected'
}

export type SwapAmount = SwapAmount_WithDesiredInput | SwapAmount_WithDesiredOutput

export interface SwapAmount_WithDesiredInput {
    __kind: 'WithDesiredInput'
    desiredAmountIn: bigint
    minAmountOut: bigint
}

export interface SwapAmount_WithDesiredOutput {
    __kind: 'WithDesiredOutput'
    desiredAmountOut: bigint
    maxAmountIn: bigint
}

export type LiquiditySourceType = LiquiditySourceType_BondingCurvePool | LiquiditySourceType_MockPool | LiquiditySourceType_MockPool2 | LiquiditySourceType_MockPool3 | LiquiditySourceType_MockPool4 | LiquiditySourceType_MulticollateralBondingCurvePool | LiquiditySourceType_XSTPool | LiquiditySourceType_XYKPool

export interface LiquiditySourceType_BondingCurvePool {
    __kind: 'BondingCurvePool'
}

export interface LiquiditySourceType_MockPool {
    __kind: 'MockPool'
}

export interface LiquiditySourceType_MockPool2 {
    __kind: 'MockPool2'
}

export interface LiquiditySourceType_MockPool3 {
    __kind: 'MockPool3'
}

export interface LiquiditySourceType_MockPool4 {
    __kind: 'MockPool4'
}

export interface LiquiditySourceType_MulticollateralBondingCurvePool {
    __kind: 'MulticollateralBondingCurvePool'
}

export interface LiquiditySourceType_XSTPool {
    __kind: 'XSTPool'
}

export interface LiquiditySourceType_XYKPool {
    __kind: 'XYKPool'
}
// ...
export const FilterMode: sts.Type<FilterMode> = sts.closedEnum(() => {
    return  {
        AllowSelected: sts.unit(),
        Disabled: sts.unit(),
        ForbidSelected: sts.unit(),
    }
})

export const LiquiditySourceType: sts.Type<LiquiditySourceType> = sts.closedEnum(() => {
    return  {
        BondingCurvePool: sts.unit(),
        MockPool: sts.unit(),
        MockPool2: sts.unit(),
        MockPool3: sts.unit(),
        MockPool4: sts.unit(),
        MulticollateralBondingCurvePool: sts.unit(),
        XSTPool: sts.unit(),
        XYKPool: sts.unit(),
    }
})

export const AssetId32: sts.Type<AssetId32> = sts.struct(() => {
    return  {
        code: sts.bytes(),
    }
})

export const SwapBatchInfo: sts.Type<SwapBatchInfo> = sts.struct(() => {
    return  {
        outcomeAssetId: AssetId32,
        outcomeAssetReuse: sts.bigint(),
        dexId: sts.number(),
        receivers: sts.array(() => BatchReceiverInfo),
    }
})

export const BatchReceiverInfo: sts.Type<BatchReceiverInfo> = sts.struct(() => {
    return  {
        accountId: AccountId32,
        targetAmount: sts.bigint(),
    }
})
// ...
export const AccountId32 = sts.bytes()